# core/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_alpkeskin_rota_core_internal_models.AuthenticationSettings:
    properties:
      enabled:
        description: Enable authentication for proxy requests
        type: boolean
      password:
        description: Password for proxy authentication (write-only)
        type: string
      username:
        description: Username for proxy authentication
        type: string
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
    type: object
  github_com_alpkeskin_rota_core_internal_models.HealthCheckSettings:
    properties:
      headers:
        items:
          type: string
        type: array
      status:
        type: integer
      timeout:
        type: integer
      url:
        type: string
      workers:
        type: integer
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.LogRetentionSettings:
    properties:
      cleanup_interval_hours:
        description: How often to run cleanup (1, 6, 12, 24)
        type: integer
      compression_after_days:
        description: Compress logs older than X days (1, 3, 7, 14)
        type: integer
      enabled:
        description: Enable automatic log cleanup
        type: boolean
      retention_days:
        description: Days to keep logs (7, 15, 30, 60, 90)
        type: integer
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.RateLimitSettings:
    properties:
      enabled:
        type: boolean
      interval:
        description: in seconds
        type: integer
      max_requests:
        type: integer
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.RotationSettings:
    properties:
      allowed_protocols:
        description: '["http", "https", "socks4", "socks4a", "socks5"], empty means
          all'
        items:
          type: string
        type: array
      fallback:
        type: boolean
      fallback_max_retries:
        type: integer
      follow_redirect:
        type: boolean
      max_response_time:
        description: in milliseconds, 0 means no limit
        type: integer
      method:
        type: string
      min_success_rate:
        description: 0-100, 0 means no minimum
        type: number
      remove_unhealthy:
        type: boolean
      retries:
        type: integer
      time_based:
        $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.TimeBasedSettings'
      timeout:
        type: integer
    type: object
  github_com_alpkeskin_rota_core_internal_models.Settings:
    properties:
      authentication:
        $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.AuthenticationSettings'
      healthcheck:
        $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.HealthCheckSettings'
      log_retention:
        $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.LogRetentionSettings'
      rate_limit:
        $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.RateLimitSettings'
      rotation:
        $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.RotationSettings'
    type: object
...
  github_com_alpkeskin_rota_core_internal_models.TimeBasedSettings:
    properties:
      interval:
        description: in seconds
        type: integer
    type: object
...
host: localhost:8001
info:
  contact:
    name: API Support
    url: https://github.com/alpkeskin/rota
  description: |-
    A high-performance proxy rotation server with health monitoring and intelligent routing
    Provides comprehensive API for managing proxy servers, monitoring their health,
    and configuring rotation strategies.
  license:
    name: LICENSE
    url: https://github.com/alpkeskin/rota/blob/main/LICENSE
  title: Rota Proxy API
  version: 1.0.0
paths:
...
  /settings:
    get:
      description: Get current system configuration
      produces:
      - application/json
      responses:
        "200":
          description: Current settings
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.Settings'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Get settings
      tags:
      - settings
    put:
      consumes:
      - application/json
      description: Update system configuration
      parameters:
      - description: Updated settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.Settings'
      produces:
      - application/json
      responses:
        "200":
          description: Update confirmation
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Update settings
      tags:
      - settings
  /settings/reset:
    post:
      description: Reset configuration to default values
      produces:
      - application/json
      responses:
        "200":
          description: Reset confirmation
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_alpkeskin_rota_core_internal_models.ErrorResponse'
      summary: Reset settings
      tags:
      - settings
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"